Language.Haskell.Tools.BackendGHC

Description

The FromGHC module provides a way to transform the GHC AST into our AST. This transformation is done in the Ghc monad. The conversion can be performed from the Parsed and the Renamed GHC AST. If the renamed AST is given, additional semantic information is looked up while traversing the AST.

Documentation

trfModule :: ModSummary -> Located (HsModule GhcPs) -> Trf (Ann UModule (Dom GhcPs) RangeStage)

trfModuleRename :: ModSummary -> Ann UModule (Dom GhcPs) RangeStage -> RenamedSource -> Located (HsModule GhcPs) -> Trf (Ann UModule (Dom GhcRn) RangeStage)

Transformes the module in its typed state. Uses the results of trfModule to extract program elements (splices for example) that are not kept in the typed representation.

addTypeInfos :: LHsBinds GhcTc -> Ann UModule (Dom GhcRn) RangeStage -> Ghc (Ann UModule IdDom RangeStage)

runTrf :: Map ApiAnnKey [SrcSpan] -> Map String [Located String] -> Trf a -> Ghc a

Performs the transformation given the tokens of the source file

data SpliceInsertionProblem

Constructors

SpliceInsertionProblem SrcSpan String

data ConvertionProblem

Constructors

ConvertionProblem SrcSpan String

UnrootedConvertionProblem String