Language.Haskell.Tools.BackendGHC.Names

Description

Functions that convert the basic elements of the GHC AST to corresponding elements in the Haskell-tools AST representation

Documentation

trfOperator :: forall n r. TransformName n r => Located (IdP n) -> Trf (Ann UOperator (Dom r) RangeStage)

trfOperator' :: forall n r. TransformName n r => IdP n -> Trf (UOperator (Dom r) RangeStage)

trfName :: forall n r. TransformName n r => Located (IdP n) -> Trf (Ann UName (Dom r) RangeStage)

trfName' :: forall n r. TransformName n r => IdP n -> Trf (UName (Dom r) RangeStage)

trfAmbiguousFieldName :: TransformName n r => Located (AmbiguousFieldOcc n) -> Trf (Ann UName (Dom r) RangeStage)

trfAmbiguousFieldName' :: forall n r. TransformName n r => SrcSpan -> AmbiguousFieldOcc n -> Trf (Ann UName (Dom r) RangeStage)

trfAmbiguousOperator' :: forall n r. TransformName n r => SrcSpan -> AmbiguousFieldOcc n -> Trf (Ann UOperator (Dom r) RangeStage)

type CorrectPass n = (Data (HsLit n), Outputable (HsLit n), Data (HsType n), Outputable (HsType n), Data (Pat n), Outputable (Pat n), Data (HsExpr n), Outputable (HsExpr n), Data (Stmt n (LHsExpr n)), Outputable (Stmt n (LHsExpr n)), Data (Stmt n (LHsCmd n)), Outputable (Stmt n (LHsCmd n)), Data (HsCmd n), Outputable (HsCmd n), Data (Sig n), Outputable (Sig n), Data (HsLocalBinds n), Outputable (HsLocalBinds n), Data (HsBind n), Outputable (HsBind n), Data (IdP n), Outputable (IdP n), Data (ConDecl n), Outputable (ConDecl n), Data (HsDecl n), Outputable (HsDecl n), Data (HsSplice n), Outputable (HsSplice n))

type ConvOk n = (XExprWithTySig n ~ LHsSigWcType n, XAppTypeE n ~ LHsWcType n, NameOrRdrName (IdP n) ~ IdP n)

class (ConvOk n, Eq n, CorrectPass n, GHCName n, FromGHCName (IdP n), HasOccName (IdP n)) => TransformableName n where

Methods

correctNameString :: IdP n -> Trf String

transformSplice :: HsSplice GhcPs -> Trf (HsSplice n)

class (TransformableName name, HsHasName (IdP name), FromGHCName (IdP res), Eq (IdP name), GHCName res, NameOrRdrName (IdP name) ~ IdP name, XUnambiguous name ~ XCFieldOcc name) => TransformName name res where

This class allows us to use the same transformation code for multiple variants of the GHC AST. GHC UName annotated with name can be transformed to our representation with semantic annotations of res.

Methods

transformName :: IdP name -> IdP res

Demote a given name

trfNameText :: String -> Trf (Ann UName (Dom r) RangeStage)

trfImplicitName :: HsIPName -> Trf (Ann UName (Dom r) RangeStage)

isOperatorStr :: String -> Bool

trfQualifiedName :: forall n r. TransformName n r => Bool -> Located (IdP n) -> Trf (Ann UQualifiedName (Dom r) RangeStage)

trfQualifiedNameFocus :: forall n r. TransformName n r => Bool -> IdP n -> Trf (Ann UQualifiedName (Dom r) RangeStage)

trfQualifiedName' :: forall n r. TransformName n r => IdP n -> Trf (UQualifiedName (Dom r) RangeStage)

trfOperatorStr :: Bool -> String -> Trf (AnnListG UNamePart (Dom r) RangeStage)

trfNameStr :: Bool -> String -> Trf (AnnListG UNamePart (Dom r) RangeStage)

Creates a qualified name from a name string

trfNameStr' :: String -> SrcLoc -> [Ann UNamePart (Dom r) RangeStage]

trfFastString :: Located FastString -> Trf (Ann UStringNode (Dom r) RangeStage)